refactor(mobile/ResultsHeaderAndError): Extract component. Refactors.
